Removed running flag from timeclass to config class and added controller code stub
