+++ src/com/group33/controller/Controller.java
@@ -1,0 +1,7 @@
+package com.group33.controller;
+
+public class Controller implements Runnable{
+
+    @Override
+    public void run() {}
+}
