+++ src/com/group33/controller/Generator.java
@@ -1,0 +1,6 @@
+package com.group33.controller;
+
+public class Generator {
+
+
+}
